EDA_Iris_Dataset
During the EDA process, I explored the dataset, visualized the distribution of features, and analyzed the correlations between variables. The pairplot helped me identify patterns and relationships between the features, while the correlation heatmap provided a clear overview of the feature relationships.

One of the key findings from the EDA was the distinct differences in the sepal and petal measurements among the three iris species - setosa, versicolor, and virginica. The box plots highlighted these differences and showcased the unique characteristics of each species.